% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_text.R
\name{match_text}
\alias{match_text}
\title{Match sentiment Words in a Text}
\usage{
match_text(word_vector, sentiment_vector)
}
\arguments{
\item{word_vector}{A character vector of words from the document.}

\item{sentiment_vector}{A character vector of sentiment words including prefix ending with a '*'.}
}
\value{
An integer count of matching sentiment words.
}
\description{
This function compares a vector of words from a text document to a vector
of sentiment words (positive or negative) and counts how many matches occur.
Supports words with prefix as a sentiment word.
}
\examples{
match_text(c("happy", "joyful", "sad", "sadly"), c("happy*", "sad*"))

}
